ArmVExpressFastBoot: Add support for sparse Images
Add support for current Fastboot Sparse images. Tested to be able
to flash both a a small image (60MB) And a large 1.5GB image. The
code in it's current form doesn't do many checks, so no defences
against corrupt images.

ArmVExpressFastBoot is obviously the wrong place to implement this,
Implementing it correctly in AndroidFastbootApp requires changing
FASTBOOT_PLATFORM_FLASH with a "offset" parameter.

Cc: Zhangfei Gao <zhangfei.gao@linaro.org>
Cc: Olivier Martin <olivier.martin@arm.com>
Signed-off-by: Riku Voipio <riku.voipio@linaro.org>