linaro-edk2: Enable "adb reboot bootloader" support
This allows the bootloader to user a reserved page to
read if the kernel requested we boot into fastboot
mode or normal bootup.

This allows users to reboot into fastboot mode w/o
having to modify jumpers on J15

This change requires the kernel to support the
communication of the reboot reason to the bootloader.

Cc: Vishal Bhoj <vishal.bhoj@linaro.org>
Cc: haojian.zhuang@linaro.org
Cc: guodong.xu@linaro.org
Signed-off-by: John Stultz <john.stultz@linaro.org>